Further tweaks to watch script.
Use a selected conversation id rather than an index,
this'll maintain the selection even if the underlying list changes.

+++ watch.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 function cleanup() {
-	kill $st_pid
+	pkill slack-term-c
 	kill $tail_pid
 }
 
@@ -9,22 +9,21 @@ trap cleanup EXIT
 
 for (( ; ; ))
 do
-	./build.sh 2>&1
-	true > dbg.log
-	true > err.log
+	./build.sh
+	build_result=$?
+	if [ $build_result -eq 0 ]; then
+		true > dbg.log
+		true > err.log
 
-	kill $st_pid
-	(alacritty -e ./slack-term-c && cat err.log)&
-	st_pid=$!
+		pkill slack-term-c
+		(alacritty -e ./slack-term-c && cat err.log)&
 
-	kill $tail_pid
-	tail -f ./dbg.log&
-	tail_pid=$!
+		kill $tail_pid
+		tail -f ./dbg.log&
+		tail_pid=$!
+	fi
 
 	# Wait for file changes, clear the terminal, and build
 	inotifywait -q -r -e create,modify,move,delete *.h *.c && \
 	  echo -ne "\033c"
 done
-
-# Re-exec ourselves for the next edit
-exec $0 $@
